pyproject.toml
[tool.poetry]
name = "llm-dataset-digest"
version = "0.0.1"
description = "Llm Dataset Digest"
authors = ["Fangning Zheng <Fangning.Zheng@tgs.com>"]
license = "TGS"
readme = "README.md"
homepage = "https://dev.azure.com/TGSCloud/Datascience/_git/llm-dataset-digest"
repository = "https://dev.azure.com/TGSCloud/Datascience/_git/llm-dataset-digest"
documentation = "https://dsdocs.cloud.tgs.com/llm-dataset-digest"
classifiers = [
  "Development Status :: 1 - Planning",
  "Programming Language :: Python",
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.5"
debugpy = "1.6.7"
streamlit = "^1.25.0"
langchain = "0.0.264"
streamlit-extras = "^0.3.0"
langchain-experimental = "^0.0.9"
openai = "^0.27.8"
faiss-cpu = "^1.7.4"
tiktoken = "^0.4.0"
country-named-entity-recognition = "^0.4"
python-pptx = "^0.6.21"
cruft = "^2.15.0"

[tool.poetry.group.dev.dependencies]
tox = "^4.6.4"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
xdoctest = {version = "^1.1.1", extras = ["all"]}

[tool.poetry.group.type.dependencies]
mypy = "^1.4.1"

[tool.poetry.group.lint.dependencies]
black = {version = "^23.7.0", extras = ["jupyter"]}
pre-commit = "^3.3.3"
ruff = "^0.0.278"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.18"
mkdocs-click = "^0.8.0"
mkdocstrings = {version = "^0.22.0", extras = ["python"]}

[tool.poetry.scripts]
llm-dataset-digest = "llm_dataset_digest.__main__:main"

[tool.ruff]
src = ["src"]
select = [
    "E",  # pycodestyle
    "F",   # pyflakes
    "B",  # bugbear
    "I",  # isort
    "UP",  # pyupgrade
    "N",  # pep8-naming
    "D",   # pydocstyle
    "ANN",  # annotations
    "S",  # bandit
    "A",  # builtins
    "C4",  # comprehensions
    "DTZ",  # datetimez
    "EM",  # errmsg
    "ICN",  # import-conventions
    "PIE",  # pie
    "PT",  # pytest-style
    "RSE",  # raise
    "RET",  # return
    "SIM",  # simplify
    "TID",  # tidy-imports
    "TCH",  # type-checking
    "ARG",  # unused-arguments
    "PTH",  # use-pathlib
    "TD",  # todos
    "PL",  # pylint
    "FLY", # flynt
    "NPY",  # numpy
]
ignore = [
    "ANN101",  # Missing type annotation for `self`
]
line-length = 88

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.flake8-annotations]
mypy-init-return = true

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
force-single-line = true

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["llm_dataset_digest", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
